# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContractStatus = typeof ContractStatus[keyof typeof ContractStatus];


export const ContractStatus = {
  pending: 'pending',
  signed: 'signed',
  expired: 'expired',
} as const;

export interface Contract {
  id: number;
  title: string;
  signerEmail: string;
  signerName: string;
  /** @nullable */
  signerCompany?: string | null;
  status: ContractStatus;
  /** @nullable */
  originalFileUrl?: string | null;
  /** @nullable */
  signedPdfUrl?: string | null;
  token: string;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  signerIp?: string | null;
  /** @nullable */
  signerLocation?: string | null;
  /** @nullable */
  signerDevice?: string | null;
  /** @nullable */
  selfieUrl?: string | null;
  /** @nullable */
  signedAt?: string | null;
  createdAt: string;
  /** @nullable */
  customFields?: string | null;
}

export interface ContractInput {
  /** @minLength 1 */
  title: string;
  signerEmail: string;
  signerName: string;
  signerCompany?: string;
  originalFileUrl?: string;
  originalFileName?: string;
  customFields?: string;
  expiresInDays?: number;
}

export interface SigningLinkInput {
  expiresInDays?: number;
}

export interface SigningLinkResult {
  signingUrl: string;
  token: string;
  expiresAt: string;
}

export interface SigningPageData {
  contractId: number;
  title: string;
  signerName: string;
  signerEmail: string;
  /** @nullable */
  signerCompany?: string | null;
  /** @nullable */
  originalFileUrl?: string | null;
  /** @nullable */
  customFields?: string | null;
  /** @nullable */
  expiresAt?: string | null;
}

export interface SignatureSubmission {
  signerName: string;
  signerEmail: string;
  signerCompany?: string;
  signerAddress?: string;
  signatureDataUrl: string;
  selfieDataUrl?: string;
  ipAddress?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  locationString?: string;
  deviceInfo?: string;
  timestamp?: string;
}

export interface SignatureResult {
  success: boolean;
  signedPdfUrl: string;
  contractId: number;
  message?: string;
}

export interface DownloadResult {
  downloadUrl: string;
  filename: string;
}

export interface DashboardStats {
  totalContracts: number;
  pendingContracts: number;
  signedContracts: number;
  expiredContracts: number;
  signedThisMonth: number;
  /** @nullable */
  avgSigningTimeHours: number | null;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  contract_created: 'contract_created',
  contract_signed: 'contract_signed',
  contract_expired: 'contract_expired',
  link_generated: 'link_generated',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  contractTitle: string;
  signerName: string;
  timestamp: string;
}

export interface Settings {
  id: number;
  businessName: string;
  businessEmail: string;
  /** @nullable */
  businessPhone?: string | null;
  /** @nullable */
  businessAddress?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccount?: string | null;
  /** @nullable */
  bankIfsc?: string | null;
  /** @nullable */
  bankSwift?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  paypalEmail?: string | null;
  /** @nullable */
  wiseEmail?: string | null;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  panNumber?: string | null;
  /** @nullable */
  notifyEmail?: string | null;
  emailOnSign?: boolean;
}

export interface SettingsUpdate {
  businessName?: string;
  businessEmail?: string;
  businessPhone?: string;
  businessAddress?: string;
  logoUrl?: string;
  bankName?: string;
  bankAccount?: string;
  bankIfsc?: string;
  bankSwift?: string;
  upiId?: string;
  paypalEmail?: string;
  wiseEmail?: string;
  gstNumber?: string;
  panNumber?: string;
  notifyEmail?: string;
  emailOnSign?: boolean;
}

export type ListContractsParams = {
status?: ListContractsStatus;
search?: string;
};

export type ListContractsStatus = typeof ListContractsStatus[keyof typeof ListContractsStatus];


export const ListContractsStatus = {
  pending: 'pending',
  signed: 'signed',
  expired: 'expired',
  all: 'all',
} as const;