# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemType } from './activityItemType';

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  contractTitle: string;
  signerName: string;
  timestamp: string;
}