# lib/api-zod/src/generated/types/contractInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ContractInput {
  /** @minLength 1 */
  title: string;
  signerEmail: string;
  signerName: string;
  signerCompany?: string;
  originalFileUrl?: string;
  originalFileName?: string;
  customFields?: string;
  expiresInDays?: number;
}