# lib/api-zod/src/generated/types/downloadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadResult {
  downloadUrl: string;
  filename: string;
}