# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  businessName: string;
  businessEmail: string;
  /** @nullable */
  businessPhone?: string | null;
  /** @nullable */
  businessAddress?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bankName?: string | null;
  /** @nullable */
  bankAccount?: string | null;
  /** @nullable */
  bankIfsc?: string | null;
  /** @nullable */
  bankSwift?: string | null;
  /** @nullable */
  upiId?: string | null;
  /** @nullable */
  paypalEmail?: string | null;
  /** @nullable */
  wiseEmail?: string | null;
  /** @nullable */
  gstNumber?: string | null;
  /** @nullable */
  panNumber?: string | null;
  /** @nullable */
  notifyEmail?: string | null;
  emailOnSign?: boolean;
}