# lib/api-zod/src/generated/types/settingsUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsUpdate {
  businessName?: string;
  businessEmail?: string;
  businessPhone?: string;
  businessAddress?: string;
  logoUrl?: string;
  bankName?: string;
  bankAccount?: string;
  bankIfsc?: string;
  bankSwift?: string;
  upiId?: string;
  paypalEmail?: string;
  wiseEmail?: string;
  gstNumber?: string;
  panNumber?: string;
  notifyEmail?: string;
  emailOnSign?: boolean;
}