# lib/api-zod/src/generated/types/signatureResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SignatureResult {
  success: boolean;
  signedPdfUrl: string;
  contractId: number;
  message?: string;
}