# lib/api-zod/src/generated/types/signatureSubmission.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Saathi Sign - Freelancer Contract & E-Signature Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface SignatureSubmission {
  signerName: string;
  signerEmail: string;
  signerCompany?: string;
  signerAddress?: string;
  signatureDataUrl: string;
  selfieDataUrl?: string;
  ipAddress?: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  locationString?: string;
  deviceInfo?: string;
  timestamp?: string;
}